requirements.txt
streamlit
openai
python-dotenv
PyPDF2
pytesseract
pdf2image
Pillow
tiktoken
faiss-cpu
numpy
matplotlib
streamlit-extras